\name{DriftBurstHypothesis-package}
\alias{DriftBurstHypothesis-package}
\alias{DriftBurstHypothesis}
\docType{package}
\title{
\packageTitle{DriftBurstHypothesis}
}
\description{
\packageDescription{DriftBurstHypothesis}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{DriftBurstHypothesis}
\packageIndices{DriftBurstHypothesis}
}
\author{
\packageAuthor{DriftBurstHypothesis}

Maintainer: \packageMaintainer{DriftBurstHypothesis}
}
\references{
Christensen, Oomen and Reno (2018) <DOI:10.2139/ssrn.2842535>.
}

\keyword{ Drift burst hypothesis }


